third_party/blink/web_tests/platform/generic/fast/dom/Document/CaretRangeFromPoint/basic-expected.txt

Testing upper left
PASS: range.startContainer == element.firstChild.
PASS: range.startOffset == 0.
Testing upper right
PASS: range.startContainer == element.firstChild.
PASS: range.startOffset == 23.
Testing lower left
PASS: range.startContainer == element.firstChild.
PASS: range.startOffset == 72.
Testing lower right
PASS: range.startContainer == element.firstChild.
PASS: range.startOffset == 95.
Testing somewhere in the middle
PASS: range.startContainer == element.firstChild.
PASS: range.startOffset == 36.
Testing negative values
PASS: range is null.
PASS: range is null.
PASS: range is null.
Testing values larger than the viewport
PASS: range is null.
PASS: range is null.
PASS: range is null.
